refactor: do not use WAL for transfer
